requirements.txt
ale-py==0.7.4
annotated-types==0.7.0
antlr4-python3-runtime==4.8
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
configparser==7.1.0
docker-pycreds==0.4.0
einops==0.3.2
eval_type_backport==0.2.2
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
gym==0.25.1
gym-notices==0.0.8
hydra-core==1.1.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.1
#nvidia-cublas-cu11==11.11.3.6
#nvidia-cuda-cupti-cu11==11.8.87
#nvidia-cuda-nvrtc-cu11==11.8.89
#nvidia-cuda-runtime-cu11==11.8.89
#nvidia-cudnn-cu11==9.1.0.70
#nvidia-cufft-cu11==10.9.0.58
#nvidia-curand-cu11==10.3.0.86
#nvidia-cusolver-cu11==11.4.1.48
#nvidia-cusparse-cu11==11.7.5.86
#nvidia-nccl-cu11==2.21.5
#nvidia-nvtx-cu11==11.8.86
omegaconf==2.1.2
opencv-python==4.11.0.86
pathtools==0.1.2
pillow==11.0.0
platformdirs==4.3.6
promise==2.3
protobuf==3.20.3
psutil==5.8.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.20.0
setproctitle==1.3.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
subprocess32==3.5.4
sympy==1.13.1
termcolor==2.3.0
torch
torchaudio
torchvision
tqdm==4.62.3
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.5
yaspin==3.1.0
zipp==3.21.0